requirements.txt
discord.py==1.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
oauth2client==4.1.3
oauthlib==3.1.0
